Fix race-condition during multi-byte reads
Bad readings can occur when multi-byte reads happen when the lsb is
close to under/overflow, e.g. when the value is teetering between 255
and 256, if the change happens between the two readings, msb and lsb
can be out-of-sync, so we get 0 instead of 256.

This change fixes the above, by reading both bytes in the same call.